+++ programs/__init__.py
@@ -1,7 +1,7 @@
 """
 Import ararpy
 """
-version = "20250523"
+version = "20250604"
 # import ararpy as ap                 # using Pip version
 try:
     from .local_init import *